change of plans
Apparently we are still making a platform fighter, even though the team lead approved of the beat em up style.
here is the skeleton for a platformer. we can build off of this.
the beat em up scene is still a part of the project until we have 100% confirmed that we will not be using that format. useful if we ever change our minds